Firewallhausmeister/requirements.txt
Pillow~=9.2.0
pywin32==304